refactor(factory): reuse the shared useNow ticker, drop the duplicate
Follow-up on prix-cloud's review of #840: the extraction introduced a second
useNow in hooks/ when mission-control/useNow.ts already had an optimized shared
ticker (one setInterval per interval bucket, bucketed for referential stability).
Point TerminalDetail + UnifiedAgentsPane at the existing mission-control/useNow
and delete hooks/useNow.ts, so every Floor consumer shares one timer instead of
holding its own.

271 mission-control tests pass; tsc + vite build clean.
