+++ apps/factory/ui/settings/components/mission-control/TerminalDetail.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 import type { TerminalDetail as TerminalInfo } from '../../types'
-import { postMessage, useNow } from '../../hooks'
+import { postMessage } from '../../hooks'
 import { renderTodoDescription, renderMarkdown } from '../../utils/markdown'
+import { useNow } from './useNow'
 import { VerticalTimeline } from './Timeline'
 import { TodoChecklist } from './TodoChecklist'
 import { latestTodos } from './floorModel'
