+++ apps/factory/ui/settings/hooks/index.ts
@@ -1,4 +1,3 @@
 export * from './useTheme'
 export * from './useVscodeApi'
 export * from './usePanelVisibility'
-export * from './useNow'
